+++ common_file_reader.h
@@ -1,22 +1,21 @@
-//
-// Created by matias on 15/9/21.
-//
-
 #ifndef FILE_READER_H
 #define FILE_READER_H
 
 #include <stdio.h>
 #include <string.h>
+#include <stdbool.h>
+#include <stdlib.h>
+#include <stdint.h>
 
 typedef struct {
     FILE * fds;
 } FileReader;
 
-int fileReaderInit(FileReader * self, FILE * fds);
+uint16_t fileReaderInit(FileReader * self, FILE * fds);
 void fileReaderUnInit(FileReader * self);
-int fileReaderInitFromName(FileReader * self, char * name);
+uint16_t fileReaderInitFromName(FileReader * self, char * name);
 bool fileReaderEOF(FileReader * self);
-int fileReaderReadLine(FileReader * self, char * output, size_t size);
+ssize_t fileReaderReadLine(FileReader * self, char * output, size_t size);
 
 
 #endif
